+++ gms-server/src/main/java/org/gms/client/processor/npc/FredrickProcessor.java
@@ -78,7 +78,7 @@ public class FredrickProcessor {
             }
         }
 
-        int netMeso = chr.getMerchantNetMeso();
+        long netMeso = chr.getMerchantNetMeso();
         if (netMeso > 0) {
             if (!chr.canHoldMeso(netMeso)) {
                 return 0x1F;
