+++ gms-server/src/main/java/org/gms/client/processor/npc/StorageProcessor.java
@@ -198,9 +198,9 @@ public class StorageProcessor {
                     c.sendPacket(PacketCreator.enableActions());
                     break;
                 case 7: { // Mesos
-                    int meso = p.readInt();
-                    int storageMesos = storage.getMeso();
-                    int playerMesos = chr.getMeso();
+                    long meso = p.readLong();
+                    long storageMesos = storage.getMeso();
+                    long playerMesos = chr.getMeso();
 
                     if (hasGMRestrictions(chr)) {
                         chr.dropMessage(1, gmBlockedStorageMessage);
@@ -211,13 +211,13 @@ public class StorageProcessor {
 
                     if ((meso > 0 && storageMesos >= meso) || (meso < 0 && playerMesos >= -meso)) {
                         if (meso < 0 && (storageMesos - meso) < 0) {
-                            meso = Integer.MIN_VALUE + storageMesos;
+                            meso = Long.MIN_VALUE + storageMesos;
                             if (meso < playerMesos) {
                                 c.sendPacket(PacketCreator.enableActions());
                                 return;
                             }
                         } else if (meso > 0 && (playerMesos + meso) < 0) {
-                            meso = Integer.MAX_VALUE - playerMesos;
+                            meso = Long.MAX_VALUE - playerMesos;
                             if (meso > storageMesos) {
                                 c.sendPacket(PacketCreator.enableActions());
                                 return;
