+++ gms-server/src/main/java/org/gms/scripting/npc/NPCConversationManager.java
@@ -305,7 +305,7 @@ public class NPCConversationManager extends AbstractPlayerInteraction {
         return completeQuest(id, npc);
     }
 
-    public int getMeso() {
+    public long getMeso() {
         return getPlayer().getMeso();
     }
